integrate the Renew process into a function.
