+++ controllers/containerTrade.js
@@ -7,11 +7,15 @@ const debug = require('../helpers/debugger')('containerTrade');
 const DEMO_CONTAINER_ID_LIST = require('../config/config').demoContainers;
 const validateStateChanging = require('../helpers/toolkit').validateStateChanging;
 const bindFunction = require('../helpers/toolkit').bindFunction;
+const dateCheckpoint=require('../helpers/toolkit').dateCheckpoint;
+const intReLength=require('../helpers/toolkit').intReLength;
+const tradeCallback=require('../controllers/tradeCallback');
 const DataCacheFactory = require("../models/dataCacheFactory");
 const Container = require('../models/DB/containerDB');
 const Trade = require('../models/DB/tradeDB');
 const User = require('../models/DB/userDB');
 const Box = require('../models/DB/boxDB');
+const redis=require('../models/redis');
 
 const status = ['delivering', 'readyToUse', 'rented', 'returned', 'notClean', 'boxed'];
 const REAL_ID_RANGE = 99900;
@@ -302,4 +306,179 @@ function getOriUser(action, theContainer, rentToUser, cb) {
     }, cb);
 }
 
-module.exports = changeContainersState;
+function Renew_Containers(dbStore,containers,storeID,boxID,orderTime){
+    return new Promise(function(resolve,reject){
+        new Promise(function(resolve,reject){
+            changeContainersState(
+                containers,
+                dbStore, {
+                    action: 'Return',
+                    newState: 3
+                }, {
+                    storeID,
+                    orderTime:orderTime,
+                    activity: "沒活動"
+                },
+                (err, tradeSuccess, reply, tradeDetail) => {
+                    if (err) reject({
+                        err:err
+                    });
+                    if (!tradeSuccess) reject({
+                        reply:reply
+                    });
+                    tradeCallback.return(tradeDetail, {
+                        storeID
+                    });
+                    let Return_reply=reply;
+                    resolve(Return_reply)
+                }
+            );
+        }).then(Return_reply=>{
+            new Promise(function(resolve,reject){
+                changeContainersState(
+                    containers,
+                    dbStore, {
+                        action: 'ReadyToClean',
+                        newState: 4
+                    }, {
+                        orderTime:orderTime,
+                    },
+                    (err, tradeSuccess, reply) => {
+                        if (err) reject({
+                            err:err
+                        });
+                        if (!tradeSuccess) reject({
+                            reply:reply
+                        });
+                        resolve(Return_reply)
+                    }
+                );
+            }).
+            then(Return_reply=>{
+                new Promise(function(resolve,reject){
+                    let task=function(done){
+                        changeContainersState(
+                            containers,
+                            dbStore, {
+                                action: 'Boxing',
+                                newState: 5
+                            }, {
+                                boxID,
+                            },
+                            (err, tradeSuccess, reply) => {
+                                if (err) reject({
+                                    err:err
+                                });
+                                if (!tradeSuccess) reject({
+                                    reply:reply
+                                });
+                                done()
+                            }
+                        );
+                    }
+                    redis.get('boxCtr', (err, boxCtr) => {
+                        if (err) reject({
+                            err:err
+                        });
+                        if (boxCtr == null) boxCtr = 1;
+                        else boxCtr++;
+                        redis.set('boxCtr', boxCtr, (err, reply) => {
+                            if (err) reject({
+                                err:err
+                            });
+                            if (reply !== 'OK') reject({
+                                reply:reply
+                            });
+                            redis.expire(
+                                'boxCtr',
+                                Math.floor((dateCheckpoint(1).valueOf() - Date.now()) / 1000),
+                                (err, reply) => {
+                                    if (err) reject({
+                                        err:err
+                                    });
+                                    if (reply !== 1) reject({
+                                        reply:reply
+                                    });
+                                    var today = new Date();
+                                    boxID =
+                                        today.getMonth() +
+                                        1 +
+                                        intReLength(today.getDate(), 2) +
+                                        intReLength(boxCtr, 3);
+                                    task(()=>{
+                                        resolve(Return_reply)
+                                    })
+                                }
+                            );
+                        });
+                    });
+                }).then((Return_reply)=>{
+                    new Promise(function(resolve,reject){
+                        changeContainersState(
+                            containers,
+                            dbStore, {
+                                action: 'Delivery',
+                                newState: 0,
+                            }, {
+                                boxID,
+                                storeID,
+                            },
+                            (err, tradeSuccess, reply) => {
+                                if (err) reject({
+                                    err:err
+                                });
+                                if (!tradeSuccess) reject({
+                                    reply:reply
+                                });
+                                resolve(Return_reply)
+                            }
+                        );
+                    }).then((Return_reply)=>{
+                        new Promise(function(resolve,reject){
+                            changeContainersState(
+                                containers,
+                                dbStore, {
+                                    action: 'Sign',
+                                    newState: 1
+                                }, {
+                                    boxID,
+                                    storeID: storeID,
+                                },
+                                (err, tradeSuccess, reply) => {
+                                    if (err) reject({
+                                        err:err
+                                    });
+                                    if (!tradeSuccess) reject({
+                                        reply:reply
+                                    });
+                                    resolve(Return_reply)
+                                }
+                            );
+                        }).then(Return_reply=>{
+                            resolve(Return_reply);
+                        }).catch(err_messenge=>{
+                            if(err_messenge.err) reject(err_messenge)
+                            if(err_messenge.reply) reject(err_messenge)
+                        })
+                    }).catch(err_messenge=>{
+                        if(err_messenge.err) reject(err_messenge)
+                        if(err_messenge.reply) reject(err_messenge)
+                    })
+                }).catch(err_messenge=>{
+                    if(err_messenge.err) reject(err_messenge)
+                    if(err_messenge.reply) reject(err_messenge)
+                })
+            }).catch(err_messenge=>{
+                if(err_messenge.err) reject(err_messenge)
+                if(err_messenge.reply) reject(err_messenge)
+            })
+        }).catch(err_messenge=>{
+            if(err_messenge.err) reject(err_messenge)
+            if(err_messenge.reply) reject(err_messenge)
+        })
+    })
+}
+module.exports = {
+    changeContainersState,
+    Renew_Containers
+};
